+++ src/components/BookDetails.jsx
@@ -2,8 +2,10 @@ import React, { useState } from "react";
 import AddBookButton from "./ui/AddBookButton";
 import axios from "axios";
 import { useNavigate } from 'react-router-dom';
-import { ReactNotifications, Store } from 'react-notifications-component';
+import { ReactNotifications } from 'react-notifications-component';
 import 'react-notifications-component/dist/theme.css';
+import handleAddToWishlist from '../util/WishlistFunctions';
+import Button from "@mui/material/Button";
 
 const BookDetails = (clickedBook) => {
 
@@ -36,37 +38,6 @@ const BookDetails = (clickedBook) => {
         return book;
       }
     
-    const handleNotification = (message) => {
-        Store.addNotification({
-          title: 'Success!',
-          message: message,
-          type: 'success',
-          insert: 'top',
-          container: 'top-right',
-          animationIn: ["animated", "fadeIn"],
-          animationOut: ["animated", "fadeOut"],
-          dismiss: {
-            duration: 3000,
-            onScreen: true
-          }
-        });
-    }
-
-
-
-    const handleAddToWishlist = () => {
-      const bookData = {bookKey : book.bookKey, title : book.title, bookCover : book.bookCover }
-      axios.post(
-        `http://localhost:8080/wishlist/addToWishlist`, 
-        bookData, 
-        {
-          headers: {'Authorization': `${authHeader}`}
-        }
-      )
-      .then (handleNotification( `${book.title} added to your wishlist` ));
-    }
-
-      
    const cleanBook = processCleanBook(book);
    console.log(book);
    console.log(cleanBook);
@@ -104,8 +75,8 @@ const BookDetails = (clickedBook) => {
                             </div>
                             </div>
                             <div className="book-details">
-                            <button className = "second-btn" onClick={handleAddToWishlist}>Add to Wishlist </button>
                             <AddBookButton book={cleanBook}/>
+                            <Button id="wishlist-button" variant='contained' color='success'onClick={() => handleAddToWishlist(book)}>Add to Wishlist</Button>
                             </div>
           </div>
         </>
