+++ src/components/BookView.jsx
@@ -1,12 +1,16 @@
 import React, { useState, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import Navigation from './Navigation';
+import Button from "@mui/material/Button";
 
 function BookView() {
+    const authHeader = localStorage.getItem('jwt');
     const {id} = useParams();
     const [book , setBook] = useState([]);
     const [error, setError] = useState(null);
     const [isLoaded, setIsLoaded] = useState(false);
+    const [wishlist, setWishlist]= useState([]);
+    const [bookList, setBookList]= useState([]);
 
     useEffect(() => {
         fetch('http://localhost:8080/api/books/'+id)
@@ -68,6 +72,15 @@ function BookView() {
                             <div className="book-details">
                             <button className = "second-btn">Add to Wishlist </button>
                             <button className = "third-btn">Add Books </button>
+                            <section>
+                                {(() => {
+                                    const bookRecord = bookList.find(book => book.bookKey === wishlistItem.bookKey && book.available === true)
+                                    if (bookRecord && bookRecord.available) {
+                                    return (<Button variant="contained" color="secondary" onClick={() => handleBorrow(bookRecord)}> Borrow </Button>);
+                                } else {
+                                    return (<Button variant="outlined" disabled>Unavailable</Button>);
+                                }})()}
+                                </section>
                             <button className= "fourth-btn"> Borrow Book</button>
                             </div>
